Applied huge patch (slightly modified) from David Chisnall
These fixes are for the FreeBSD/OpenBSD port.

+++ sope-core/EOControl/EOGenericRecord.m
@@ -64,7 +64,7 @@ static __inline__ void eoRemove(EOGenericRecord *table, id key);
 
 static EONull *null = nil;
 
-+ (int)version {
++ (NSInteger)version {
   return 2;
 }
