+++ sope-core/NGExtensions/FdExt.subproj/NSProcessInfo+misc.m
@@ -68,7 +68,7 @@
   NSString *s;
   cnt++;
   s = [NSString stringWithFormat:
-                  @"%04X%X%02X.tmp", getpid(), time(NULL),
+                  @"%04X%X%02X.tmp", getpid(), (unsigned int)time(NULL),
                   cnt];
   return [_prefix stringByAppendingString:s];
 }
