+++ sope-xml/SaxObjC/SaxContentHandler.h
@@ -50,7 +50,7 @@
 
 /* CDATA */
 
-- (void)characters:(unichar *)_chars          length:(int)_len;
+- (void)characters:(unichar *)_chars          length:(NSUInteger)_len;
 - (void)ignorableWhitespace:(unichar *)_chars length:(int)_len;
 
 /* PIs */
