+++ sope-xml/SaxObjC/SaxDefaultHandler.m
@@ -48,7 +48,7 @@
 {
 }
 
-- (void)characters:(unichar *)_chars length:(int)_len {
+- (void)characters:(unichar *)_chars length:(NSUInteger)_len {
 }
 - (void)ignorableWhitespace:(unichar *)_chars length:(int)_len {
 }
